Extract Types::Base#to_compact_hash to Compactable
